Pass multiple paths to a formatter
This reduces the number of exec calls to the formatters

If we have too many paths it's possible to hit ARG_MAX (though unlikely
on modern Linux systems).
We could chunk the paths into a more reasonable size if this becomes an
issue.
See: https://www.in-ulm.de/~mascheck/various/argmax/