nix: use naerks, remove complexity
Introduce naersk so we don't have to update the cargoSha256 all the
time.

Use nixpkgs rust instead of rust-overlay to make the integration with
naersk easier.

Remove the overlay as it wasn't very useful. Since we depend on
third-party dependencies, the overlay cannot be used in standalone
anyways.

Update flake.lock