move treefmt.withConfig to a separate repo (#204)
https://github.com/numtide/treefmt-nix now contains both the treefmt Nix
module configuration, and also a collection of pre-configured
formatters.

By moving that functionality over there, we can make use of the same
mechanisms on the treefmt package that is shipped with nixpkgs. Another
possibility might have been to add withConfig to nixpkgs, but from
experience nix code that is duplicated in nixpkgs starts evolving on its
own.