fix: stop symlinked tree root being deref'd
As reported in numtide/treefmt#221 by @srhb there is currently an issue
where if the `--tree-root` is a symlink everything under that path will
be ignored. This is due to the symlink being dereferenced and the
symlink path not overlapping with the target path. This commit attempts
to rectify this issue by using the `$PWD` environment variable which
in *most* shells will reflect the symlink path rather than the target.
If this environment variable cannot be read then it falls back to use
`getcwd` which will perform the present dereferencing behaviour.

Thanks to @srhb for reporting this issue and thanks to @DBCDK for
sponsoring the time to get this investigated and fixed.