Clarify in the spec that the formatter MUST process the given files
Some formatters are aware of VCS systems like git, and "helpfully" skip
over files that are not tracked by the VCS system. This doesn't play
nicely with treefmt, because treefmt expects to be able to generate
tempfiles (which are not tracked by a VCS) and pass them to the
formatter.