feat: add git staged reader
There are now two ways of traversing a git repository:

- `git_index` which uses `git ls-files`
- `git_staged` which uses `diff --name-only --cached`

The `git` walk type is an alias for `git_index` and is the default.

`git_staged` can be used to only format staged changes.

This is particularly useful when introducing `treefmt` to an existing repository, allowing the user to incrementally reformat.