feat: improve change detection
Previously, we were storing the last `modtime` and `size` for a given path in `boltdb`.
This was used to determine if a file had changed and whether we should format it.

In addition, if a formatter's executable changed (`modtime` or `size`), we would delete
all path entries in the database before processing, thereby forcing each file to be
formatted.

This commit introduces a new approach to change detection, one which takes into account
changes to the underlying file, the formatter's executable, _and_ the formatter's
configuration.

Now, when deciding if we should format a file, we do the following:

- Hash each matching formatter, in sequence, using its `name`, `options`, `priority`
  as well as the `modtime` and `size` of its executable.
  This is pre-computed on a per-pipeline basis.
- We then add the file's `modtime` and `size` to generate a format signature.
  You can think of this signature as a unique representation of what we are about to do
  with the file.
- The format signature is then compared with a cache entry (if available).
- If the signatures match, we have already applied this sequence of formatters, with these
  particular options etc. to this file when it had this `modtime` and `size`, so there is
  no more processing to be done.
- If the signatures do no match, we should format the file, recording the new format
  signature in the cache when we are finished.

This approach is simpler in terms of storage, and has the added benefit of finer grained
change detection versus the brute force cache busting we were doing before.

In terms of performance impact, with the pre-computing of hashes per-pipeline and the
simpler storage schema, there appears to have been no significant impact.
Manual testing with [nixpkgs](https://github.com/nixos/nixpkgs) shows comparable run times
for both hot and cold caches.

> [!NOTE]
> Since this changes the database schema, rather than implementing some form of migration
> logic to remove the old buckets and so on, I decided to upgrade the hash algorithm we
> use when determining the filename for the db file.
>
> Previously, we were using `sha1`, matching the behaviour from `1.0`.
> Now we use `sha256`, which results in a slightly longer db name, but has the benefit of
> ensuring a new db instance will be created on first invocation, as well as making
> [gosec](https://golangci-lint.run/usage/linters/#gosec) happy.

Closes #455

Signed-off-by: Brian McGee <brian@bmcgee.ie>