chore: fix flaky tests
I found that `TestGit` (and `TestJujutsu`, which was copied from
`TestGit`) were failing intermittently. This took me some time to
track down: the issue is an interaction between a few things:

- This test configures a formatter to run on all files (`"*"`), which
  importantly includes the `treefmt.toml` file at the root of the temp
  test project
- This test invokes treefmt with `withConfig(configPath, cfg)`. Under
  the hood, `withConfig` always writes to the `treefmt.toml` config
  file, which (possibly) means updating the mtime of `treefmt.toml` file.
- treefmt computes cache signatures using mtimes at 1-second granularity
  (presumably because this the only cross filesystem compatible option).

In effect, if the test ran quickly (under 1 second), then `treefmt.toml`
would get formatted exactly 1 time on the first invocation of treefmt.
However, if the test ran slow enough that the mtime actually changed
throughout the test, then `treefmt.toml` would get formatted multiple
times.

I could have fixed this in a few ways. I opted to remove the unnecessary
calls to `withConfig(configPath, cfg)` sprinkled throughout the test.
Furthermore, to make this less likely to recur, I changed `WriteConfig`
to ensure that the resulting file *always* has a new mtime, even if this
requires setting it to the future.